Arrange so model's XYZ system aligns with world.
Now, the forward direction for hovercraft is positive Y, side direction
is X axis, and up/down is still Z. This might be easier to reason
about/more consistent, and makes camera calculations easier. We'll see
if it's the best in the long run.

Previously, the model forward direction was positive X, and side
direction was Y axis.

Also be more consistent about order of R, Pitch, Roll.