Adjust constants to match original Hover values.
This helps the physics feel much more like the original game, which
is the goal. There is additional work to completely recreate the
original physics, but having the correct constant values is a
prerequisite for that.

The correct constant values were extracted from the production
hover.exe by @Andoryuuta. Huge thanks for this!

Also re-arrange constants to be more similar to C source; this will
make future additions easier.